i41: Move /fonts, /img, and /svg to /src/res
