+++ src/library/components/fab.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { TouchableOpacity, Text, StyleSheet } from 'react-native';
 import { typography, colors, shapes } from '../../../frontend/styles'
-import Icon from '../../../frontend/assets/svg'
+import Icon from '../../res/svg'
 
 const styles = StyleSheet.create({
