+++ src/res/img/user_marker.png
