+++ src/res/svg/ic_crosshairs_24dp.js
