+++ src/res/svg/ic_megaphone_24dp.js
