+++ src/res/svg/ic_placeholder_24dp.js
