+++ src/res/svg/ic_twitter_24dp.js
