+++ src/res/svg/index.js
